Finished PhoenixBolt -- now supports multiple inputs from multiple components. If no fields are specified, will generate upsert based on fields in first tuple from a stream. Added local cluster mode to topology definition

+++ phoenix_ddl/mtr.sql
@@ -3,10 +3,10 @@ drop table if exists mtr.mtr;
 create table mtr.mtr(
   source_ip varchar not null,
   target_ip varchar not null,
-  local_time integer not null,
+  local_time bigint not null,
   loss double,
   avg_latency double,
-  stddev_latency double 
+  stddev_latency double
   constraint my_pk primary key (source_ip, target_ip, local_time)
 ) ttl=432000
 ;
